Adding new files
new files with lower case i in index.html and in internal links. For some reason Simplifier is still exporting with an upper case I, and even though I changed the links to use i, all of the links still seem to work, so I just changed the name of the main file to index.html and left the others as is.